torch default type float64

+++ ivy/functional/backends/torch/linear_algebra.py
@@ -76,7 +76,7 @@ def matrix_norm(x: torch.Tensor,
 
 def matrix_power(x: torch.Tensor, n: int)\
         -> torch.Tensor:
-    return torch.linalg.matrix_power(x.type(torch.float64), n).type(x.dtype)
+    return torch.linalg.matrix_power(x, n)
 
 
 # noinspection PyPep8Naming
